[refactor] Domain Layer가 어느 Layer에도 의존하지 않도록 변경
