+++ app/src/main/java/com/bodan/maplecalendar/domain/entity/Status.kt
@@ -1,4 +1,4 @@
-package com.bodan.maplecalendar.presentation.utils
+package com.bodan.maplecalendar.domain.entity
 
 enum class Status {
     SUCCESS, ERROR, FAIL, LOADING
