+++ app/src/main/java/com/bodan/maplecalendar/domain/usecase/GetCharacterItemEquipmentUseCase.kt
@@ -2,7 +2,7 @@ package com.bodan.maplecalendar.domain.usecase
 
 import com.bodan.maplecalendar.domain.entity.CharacterItemEquipment
 import com.bodan.maplecalendar.domain.repository.MaplestoryRepository
-import com.bodan.maplecalendar.presentation.utils.Result
+import com.bodan.maplecalendar.domain.entity.Result
 import javax.inject.Inject
 
 class GetCharacterItemEquipmentUseCase @Inject constructor(
