+++ app/src/main/java/com/bodan/maplecalendar/domain/usecase/GetCharacterPopularityUseCase.kt
@@ -2,7 +2,7 @@ package com.bodan.maplecalendar.domain.usecase
 
 import com.bodan.maplecalendar.domain.entity.CharacterPopularity
 import com.bodan.maplecalendar.domain.repository.MaplestoryRepository
-import com.bodan.maplecalendar.presentation.utils.Result
+import com.bodan.maplecalendar.domain.entity.Result
 import javax.inject.Inject
 
 class GetCharacterPopularityUseCase @Inject constructor(
